+++ packages/libmonet/CHANGELOG.md
@@ -6,6 +6,9 @@
   - [`4578d0a`](https://github.com/material-foundation/material-color-utilities/commit/4578d0a1d40f1d17f1ca6f34cc9c484bea8f0a9d): *"Refactor DynamicScheme to accept a list of source colors."*
   - [`91da30d`](https://github.com/material-foundation/material-color-utilities/commit/91da30d89e70c3dc9575ec71a1ebe8874d881f29): *"Updated background for on-fixed colors in ColorSpec2026."*
   - [`1a34bd2`](https://github.com/material-foundation/material-color-utilities/commit/1a34bd2d696a5dd669beca679546ad72209d0937): *"Adjust chroma calculation in ColorSpec2025 for yellow hues at tone 99."*
+- Added support for energy color generation via the new `EnergyColors` utility.
+  - Implementation is experimental as it comes from an unreleased version of the upstream library.
+- Added "surface effect" colors to `AndroidDynamicColors`.
 - Made `Hct` immutable to avoid issues with value equality in collections:
   - Removed `hue`, `chroma` and `tone` setters.
   - Added a new `copyWith` method.